+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Fixed issue parsing apostrophe control symbol.
+
 ## [0.5.0] - 2021-11-04
 
 ### Added